checking permissions on contacts apis

+++ ioendpoints.py
@@ -1122,8 +1122,11 @@ class CrmEngineApi(remote.Service):
                       name='contacts.delete')
     def contact_delete(self, request):
         entityKey = ndb.Key(urlsafe=request.entityKey)
-        Edge.delete_all_cascade(start_node = entityKey)
-        return message_types.VoidMessage()
+        if Node.check_permission(user_from_email,entityKey.get()):
+            Edge.delete_all_cascade(start_node = entityKey)
+            return message_types.VoidMessage()
+        else:
+            raise endpoints.UnauthorizedException('You don\'t have permissions.')
     
 
     # contacts.insertv2 api
